Some changes after reviewing diff.
