+++ app/org/sagebionetworks/bridge/json/DateUtils.java
@@ -8,6 +8,7 @@ import org.apache.commons.lang3.StringUtils;
 import org.joda.time.DateTime;
 import org.joda.time.DateTimeZone;
 import org.joda.time.LocalDate;
+import org.joda.time.LocalTime;
 import org.joda.time.Period;
 import org.joda.time.format.DateTimeFormatter;
 import org.joda.time.format.ISODateTimeFormat;
@@ -214,4 +215,12 @@ public final class DateUtils {
         }
     }
     
+    /**
+     * Convert DateTime instance to midnight of a given day, using UTC time zone.
+     * @param dateTime
+     * @return dateTime at midnight in UTC
+     */
+    public static DateTime dateTimeToUTCMidnight(DateTime dateTime) {
+        return new DateTime(dateTime, DateTimeZone.UTC).withTime(LocalTime.MIDNIGHT);
+    }
 }
